src/flat_file_ingestor/requirements.txt
# package>=version
gcsfs>=2021.10.1
pandas>=1.3.4
google-cloud-bigquery>=2.28.1
google-cloud-storage>=1.42.0 
pyarrow>=4.0.1
openpyxl